requirements.txt
# Flask 관련
Flask==3.0.0

# AI/ML 관련
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
faiss-cpu>=1.7.4

# Git 관련
GitPython>=3.1.40

# 환경 변수 관리
python-dotenv>=1.0.0

# HTTP 요청
requests>=2.31.0

# 로깅 및 유틸리티
typing-extensions>=4.8.0
